+++ src/main/java/org/woehlke/twitterwall/oodm/entities/TickerSymbol.java
@@ -4,7 +4,7 @@ import org.woehlke.twitterwall.oodm.entities.parts.AbstractTwitterObject;
 import org.woehlke.twitterwall.oodm.entities.common.DomainObjectWithTask;
 import org.woehlke.twitterwall.oodm.entities.common.DomainObjectWithUrl;
 import org.woehlke.twitterwall.oodm.entities.parts.TaskInfo;
-import org.woehlke.twitterwall.oodm.listener.TickerSymbolListener;
+import org.woehlke.twitterwall.oodm.entities.listener.TickerSymbolListener;
 
 import javax.persistence.*;
