+++ src/main/java/org/woehlke/twitterwall/oodm/entities/User.java
@@ -6,7 +6,7 @@ import org.woehlke.twitterwall.oodm.entities.common.DomainObjectWithScreenName;
 import org.woehlke.twitterwall.oodm.entities.parts.TaskInfo;
 import org.woehlke.twitterwall.oodm.entities.common.DomainObjectWithTask;
 import org.woehlke.twitterwall.oodm.entities.parts.Entities;
-import org.woehlke.twitterwall.oodm.listener.UserListener;
+import org.woehlke.twitterwall.oodm.entities.listener.UserListener;
 
 import javax.persistence.*;
 import java.util.Date;
