+++ src/main/java/org/woehlke/twitterwall/oodm/entities/listener/UrlListener.java
@@ -1,52 +1,50 @@
-package org.woehlke.twitterwall.oodm.listener;
+package org.woehlke.twitterwall.oodm.entities.listener;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.woehlke.twitterwall.oodm.entities.HashTag;
+import org.woehlke.twitterwall.oodm.entities.Url;
 
 import javax.persistence.*;
 
 /**
  * Created by tw on 03.07.17.
  */
-public class HashTagListener {
+public class UrlListener {
 
-  private static final Logger log = LoggerFactory.getLogger(HashTagListener.class);
+  private static final Logger log = LoggerFactory.getLogger(UrlListener.class);
 
   @PrePersist
-  public void onPrePersist(HashTag domainObject) {
+  public void onPrePersist(Url domainObject) {
     log.debug("try to Persist: "+domainObject.toString());
   }
 
   @PreUpdate
-  public void onPreUpdate(HashTag domainObject) {
+  public void onPreUpdate(Url domainObject) {
     log.debug("try to Update: "+domainObject.toString());
   }
 
   @PreRemove
-  public void onPreRemove(HashTag domainObject) {
+  public void onPreRemove(Url domainObject) {
     log.debug("try to Remove: "+domainObject.toString());
   }
 
   @PostPersist
-  public void onPostPersist(HashTag domainObject) {
+  public void onPostPersist(Url domainObject) {
     log.info("Persisted: "+domainObject.toString());
   }
 
   @PostUpdate
-  public void onPostUpdate(HashTag domainObject) {
+  public void onPostUpdate(Url domainObject) {
     log.info("Updated: "+domainObject.toString());
   }
 
   @PostRemove
-  public void onPostRemove(HashTag domainObject) {
+  public void onPostRemove(Url domainObject) {
     log.info("Removed: "+domainObject.toString());
   }
 
   @PostLoad
-  public void onPostLoad(HashTag domainObject) {
+  public void onPostLoad(Url domainObject) {
     log.debug("loaded: "+domainObject.toString());
   }
-
-
 }
