+++ src/test/resources/logback.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
     <include resource="org/springframework/boot/logging/logback/base.xml"/>
-    <logger name="org.woehlke.twitterwall.oodm.listener" level="DEBUG" />
+    <logger name="org.woehlke.twitterwall.oodm.entities.listener" level="DEBUG" />
     <logger name="org.woehlke.twitterwall.oodm.repositor" level="DEBUG" />
     <logger name="org.woehlke.twitterwall.oodm.service" level="DEBUG" />
     <logger name="org.woehlke.twitterwall.oodm.entities" level="DEBUG" />
